requirements.txt
pre-commit==3.7.1
black==26.3.1
flake8==7.1.0
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
asyncpg==0.29.0
python-dotenv==1.0.1
tabulate==0.9.0
openai-agents[litellm]==0.14.2
pyyaml==6.0.2
pydantic==2.12.5
redis[asyncio]==5.2.1
aio-pika==9.4.3
